Added SQLite3 support w/ caveat
SQLite3 doesn't seem to:
* Give any value for RowCount
* Place size limits on the SQL types that are used.

+++ pony-odbc/tests/exec_direct.pony
@@ -33,9 +33,15 @@ class \nodoc\ iso _TestExecDirect is UnitTest
     var stmt: ODBCSth = ODBCSth(dbc)
     h.assert_true(stmt.exec_direct("create temporary table test_exec_direct (foo integer)"))
 
-    (var bool: Bool, var cnt: USize) = stmt.result_count()
-    h.assert_true(bool)
-    h.assert_eq[USize](0, cnt)
+    var bool: Bool = true
+    var cnt: USize = 0
+
+    // Note - SQLite3 does not return the correction rowcount
+    if (dsn != "sqlitedb3") then
+      (bool, cnt) = stmt.result_count()
+      h.assert_true(bool)
+      h.assert_eq[USize](0, cnt)
+    end
 
 /*
     /* Attempt to create the same table again - this WILL fail */
@@ -61,9 +67,14 @@ class \nodoc\ iso _TestExecDirect is UnitTest
      * results.  We just use result_count() to ensure we have the correct
      * number (1)                                                         */
     h.assert_true(stmt.exec_direct("select * from test_exec_direct"))
-    (bool, cnt) = stmt.result_count()
-    h.assert_true(bool)
-    h.assert_eq[USize](1, cnt)
+
+
+    // Note - SQLite3 does not return the correction rowcount
+    if (dsn != "sqlitedb3") then
+      (bool, cnt) = stmt.result_count()
+      h.assert_true(bool)
+      h.assert_eq[USize](1, cnt)
+    end
 
     /* Close the cursor */
     h.assert_true(stmt.finish())
@@ -73,24 +84,26 @@ class \nodoc\ iso _TestExecDirect is UnitTest
     h.assert_true(stmt.exec_direct("insert into test_exec_direct values (1)"))
     h.assert_true(stmt.exec_direct("insert into test_exec_direct values (2)"))
 
-    /* Select all the rows, count should be 4 */
-    h.assert_true(stmt.exec_direct("select * from test_exec_direct"))
-    (bool, cnt) = stmt.result_count()
-    h.assert_true(bool)
-    h.assert_eq[USize](4, cnt)
-    h.assert_true(stmt.finish())
+    // Note - SQLite3 does not return the correction rowcount
+    if (dsn != "sqlitedb3") then
+      /* Select all the rows, count should be 4 */
+      h.assert_true(stmt.exec_direct("select * from test_exec_direct"))
+      (bool, cnt) = stmt.result_count()
+      h.assert_true(bool)
+      h.assert_eq[USize](4, cnt)
+      h.assert_true(stmt.finish())
 
-    /* Delete all rows from the table.  result_count() will report how
-     * many rows were affected (ie, deleted) (4)                        */
-    h.assert_true(stmt.exec_direct("delete from test_exec_direct"))
-    (bool, cnt) = stmt.result_count()
-    h.assert_true(bool)
-    h.assert_eq[USize](4, cnt)
+      /* Delete all rows from the table.  result_count() will report how
+       * many rows were affected (ie, deleted) (4)                        */
+      h.assert_true(stmt.exec_direct("delete from test_exec_direct"))
+      (bool, cnt) = stmt.result_count()
+      h.assert_true(bool)
+      h.assert_eq[USize](4, cnt)
 
-    /* Select to ensure there are no rows left */
-    h.assert_true(stmt.exec_direct("select * from test_exec_direct"))
-    (bool, cnt) = stmt.result_count()
-    h.assert_true(bool)
-    h.assert_eq[USize](0, cnt)
-    h.assert_true(stmt.finish())
-
+      /* Select to ensure there are no rows left */
+      h.assert_true(stmt.exec_direct("select * from test_exec_direct"))
+      (bool, cnt) = stmt.result_count()
+      h.assert_true(bool)
+      h.assert_eq[USize](0, cnt)
+      h.assert_true(stmt.finish())
+    end
